Add client-side stream handler
A stream handler, if provided, is called after a response (see note) has
been received and replaces all other content-serving mechanisms.

In conjunction with a response handler, it provides the minimally
necessary interface to implement a WebSocket client.

Note: Excludes responses to CONNECT and HEAD requests, as well as
redirects.