Refactor detail::random_string()
Refactor detail::random_string() to output random bytes in addition to
random alphanumeric strings, and change the interface to fill a buffer
given a char pointer and length, instead of returning a string.

Rename the new function to random_bytes().

Add an overload random_string() that implements the old behavior on top
of random_bytes().